+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - High memory usage due to having unused emoji pickers.
+- High memory usage of the setup window, even if unused.
 
 ## [0.11.0] - 2023-12-25